pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.116.1",
    "jose>=1.0.0",
    "mercadopago>=2.3.0",
    "pandas>=2.3.1",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "stripe>=12.3.0",
    "supabase>=2.16.0",
    "yagmail>=0.15.293",
]
